+++ src/main/java/com/github/hcsp/inheritance/Dog.java
@@ -1,17 +1,10 @@
 package com.github.hcsp.inheritance;
 
-public class Dog {
-    private String name;
-
-    public Dog(String name) {
-        this.name = name;
+public class Dog extends Animal{
+    public Dog (String name) {
+        super(name);
     }
-
-    public void sayMyName() {
-        System.out.println("我的名字是" + name);
-    }
-
     public void wang() {
-        System.out.println("汪" + name);
+        System.out.println("汪" + super.getName());
     }
 }
