+++ src/main/java/com/github/hcsp/inheritance/Rat.java
@@ -1,17 +1,12 @@
 package com.github.hcsp.inheritance;
 
-public class Rat {
-    private String name;
 
+public class Rat extends Animal{
     public Rat(String name) {
-        this.name = name;
+        super(name);
     }
 
-    public void sayMyName() {
-        System.out.println("我的名字是" + name);
-    }
-
-    public void zhizhi() {
-        System.out.println("吱吱" + name);
+    public void zhizhi(){
+        System.out.println("吱吱" + super.getName());
     }
 }
